// web/src/gen/tilde/management/v1/connections_pb.d.ts
export declare type ListProvidersRequest = Message<"tilde.management.v1.ListProvidersRequest"> & {
  /**
   * @generated from field: uint32 page_size = 1;
   */
  pageSize: number;

  /**
   * @generated from field: string page_token = 2;
   */
  pageToken: string;

  /**
   * @generated from field: optional string search = 3;
   */
  search?: string | undefined;
};

/**
 * Describes the message tilde.management.v1.ListProvidersRequest.
 * Use `create(ListProvidersRequestSchema)` to create a new message.
 */
export declare const ListProvidersRequestSchema: GenMessage<ListProvidersRequest>;

/**
 * @generated from message tilde.management.v1.ListProvidersResponse
 */
export declare type ListProvidersResponse = Message<"tilde.management.v1.ListProvidersResponse"> & {
  /**
   * @generated from field: repeated tilde.types.v1.Provider providers = 1;
   */
  providers: Provider[];

  /**
   * @generated from field: string next_page_token = 2;
   */
  nextPageToken: string;
};

/**
 * Describes the message tilde.management.v1.ListProvidersResponse.
 * Use `create(ListProvidersResponseSchema)` to create a new message.
 */
export declare const ListProvidersResponseSchema: GenMessage<ListProvidersResponse>;

/**
 * @generated from message tilde.management.v1.GetProviderRequest
 */
export declare type GetProviderRequest = Message<"tilde.management.v1.GetProviderRequest"> & {
  /**
   * @generated from field: string id = 1;
   */
  id: string;
};

/**
 * Describes the message tilde.management.v1.GetProviderRequest.
 * Use `create(GetProviderRequestSchema)` to create a new message.
 */
export declare const GetProviderRequestSchema: GenMessage<GetProviderRequest>;

/**
 * @generated from message tilde.management.v1.GetProviderResponse
 */
export declare type GetProviderResponse = Message<"tilde.management.v1.GetProviderResponse"> & {
  /**
   * @generated from field: tilde.types.v1.Provider provider = 1;
   */
  provider?: Provider | undefined;
};

/**
 * Describes the message tilde.management.v1.GetProviderResponse.
 * Use `create(GetProviderResponseSchema)` to create a new message.
 */
export declare const GetProviderResponseSchema: GenMessage<GetProviderResponse>;

/**
 * @generated from message tilde.management.v1.RegisterProviderRequest
 */
export declare type RegisterProviderRequest = Message<"tilde.management.v1.RegisterProviderRequest"> & {
  /**
   * @generated from field: tilde.types.v1.Provider provider = 1;
   */
  provider?: Provider | undefined;

  /**
   * @generated from field: optional string backend_token = 3;
   */
  backendToken?: string | undefined;
};

/**
 * Describes the message tilde.management.v1.RegisterProviderRequest.
 * Use `create(RegisterProviderRequestSchema)` to create a new message.
 */
export declare const RegisterProviderRequestSchema: GenMessage<RegisterProviderRequest>;

/**
 * @generated from message tilde.management.v1.RegisterProviderResponse
 */
export declare type RegisterProviderResponse = Message<"tilde.management.v1.RegisterProviderResponse"> & {
  /**
   * @generated from field: tilde.types.v1.Provider provider = 1;
   */
  provider?: Provider | undefined;
};

/**
 * Describes the message tilde.management.v1.RegisterProviderResponse.
 * Use `create(RegisterProviderResponseSchema)` to create a new message.
 */
export declare const RegisterProviderResponseSchema: GenMessage<RegisterProviderResponse>;

/**
 * @generated from message tilde.management.v1.StartConnectionRequest
 */
export declare type StartConnectionRequest = Message<"tilde.management.v1.StartConnectionRequest"> & {
  /**
   * @generated from field: string id = 1;
   */
  id: string;

  /**
   * @generated from field: string name = 2;
   */
  name: string;

  /**
   * @generated from field: string provider_id = 3;
   */
  providerId: string;

  /**
   * @generated from field: string type_id = 4;
   */
  typeId: string;

  /**
   * @generated from field: repeated tilde.types.v1.CapabilityAssignment assignments = 5;
   */
  assignments: CapabilityAssignment[];
};

/**
 * Describes the message tilde.management.v1.StartConnectionRequest.
 * Use `create(StartConnectionRequestSchema)` to create a new message.
 */
export declare const StartConnectionRequestSchema: GenMessage<StartConnectionRequest>;

/**
 * @generated from message tilde.management.v1.StartConnectionResponse
 */
export declare type StartConnectionResponse = Message<"tilde.management.v1.StartConnectionResponse"> & {
  /**
   * @generated from field: tilde.types.v1.Connection connection = 1;
   */
  connection?: Connection | undefined;

  /**
   * @generated from field: string brokering_url = 2;
   */
  brokeringUrl: string;
};

/**
 * Describes the message tilde.management.v1.StartConnectionResponse.
 * Use `create(StartConnectionResponseSchema)` to create a new message.
 */
export declare const StartConnectionResponseSchema: GenMessage<StartConnectionResponse>;

/**
 * @generated from message tilde.management.v1.GetConnectionRequest
 */
export declare type GetConnectionRequest = Message<"tilde.management.v1.GetConnectionRequest"> & {
  /**
   * @generated from field: string id = 1;
   */
  id: string;
};

/**
 * Describes the message tilde.management.v1.GetConnectionRequest.
 * Use `create(GetConnectionRequestSchema)` to create a new message.
 */
export declare const GetConnectionRequestSchema: GenMessage<GetConnectionRequest>;

/**
 * @generated from message tilde.management.v1.GetConnectionResponse
 */
export declare type GetConnectionResponse = Message<"tilde.management.v1.GetConnectionResponse"> & {
  /**
   * @generated from field: tilde.types.v1.Connection connection = 1;
   */
  connection?: Connection | undefined;
};

/**
 * Describes the message tilde.management.v1.GetConnectionResponse.
 * Use `create(GetConnectionResponseSchema)` to create a new message.
 */
export declare const GetConnectionResponseSchema: GenMessage<GetConnectionResponse>;

/**
 * @generated from message tilde.management.v1.ListConnectionsRequest
 */
export declare type ListConnectionsRequest = Message<"tilde.management.v1.ListConnectionsRequest"> & {
  /**
   * @generated from field: uint32 page_size = 1;
   */
  pageSize: number;

  /**
   * @generated from field: string page_token = 2;
   */
  pageToken: string;

  /**
   * @generated from field: optional string agent_id = 3;
   */
  agentId?: string | undefined;

  /**
   * @generated from field: bool unassigned_channel_only = 4;
   */
  unassignedChannelOnly: boolean;
};

/**
 * Describes the message tilde.management.v1.ListConnectionsRequest.
 * Use `create(ListConnectionsRequestSchema)` to create a new message.
 */
export declare const ListConnectionsRequestSchema: GenMessage<ListConnectionsRequest>;

/**
 * @generated from message tilde.management.v1.ListConnectionsResponse
 */
export declare type ListConnectionsResponse = Message<"tilde.management.v1.ListConnectionsResponse"> & {
  /**
   * @generated from field: repeated tilde.types.v1.Connection connections = 1;
   */
  connections: Connection[];

  /**
   * @generated from field: string next_page_token = 2;
   */
  nextPageToken: string;
};

/**
 * Describes the message tilde.management.v1.ListConnectionsResponse.
 * Use `create(ListConnectionsResponseSchema)` to create a new message.
 */
export declare const ListConnectionsResponseSchema: GenMessage<ListConnectionsResponse>;

/**
 * @generated from message tilde.management.v1.AssignCapabilityRequest
 */
export declare type AssignCapabilityRequest = Message<"tilde.management.v1.AssignCapabilityRequest"> & {
  /**
   * @generated from field: string connection_id = 1;
   */
  connectionId: string;

  /**
   * @generated from field: tilde.types.v1.CapabilityAssignment assignment = 2;
   */
  assignment?: CapabilityAssignment | undefined;
};

/**
 * Describes the message tilde.management.v1.AssignCapabilityRequest.
 * Use `create(AssignCapabilityRequestSchema)` to create a new message.
 */
export declare const AssignCapabilityRequestSchema: GenMessage<AssignCapabilityRequest>;

/**
 * @generated from message tilde.management.v1.AssignCapabilityResponse
 */
export declare type AssignCapabilityResponse = Message<"tilde.management.v1.AssignCapabilityResponse"> & {
  /**
   * @generated from field: tilde.types.v1.Connection connection = 1;
   */
  connection?: Connection | undefined;
};

/**
 * Describes the message tilde.management.v1.AssignCapabilityResponse.
 * Use `create(AssignCapabilityResponseSchema)` to create a new message.
 */
export declare const AssignCapabilityResponseSchema: GenMessage<AssignCapabilityResponse>;

/**
 * @generated from message tilde.management.v1.UnassignCapabilityRequest
 */
export declare type UnassignCapabilityRequest = Message<"tilde.management.v1.UnassignCapabilityRequest"> & {
  /**
   * @generated from field: string connection_id = 1;
   */
  connectionId: string;

  /**
   * @generated from field: tilde.types.v1.CapabilityAssignment assignment = 2;
   */
  assignment?: CapabilityAssignment | undefined;
};

/**
 * Describes the message tilde.management.v1.UnassignCapabilityRequest.
 * Use `create(UnassignCapabilityRequestSchema)` to create a new message.
 */
export declare const UnassignCapabilityRequestSchema: GenMessage<UnassignCapabilityRequest>;

/**
 * @generated from message tilde.management.v1.UnassignCapabilityResponse
 */
export declare type UnassignCapabilityResponse = Message<"tilde.management.v1.UnassignCapabilityResponse"> & {
  /**
   * @generated from field: tilde.types.v1.Connection connection = 1;
   */
  connection?: Connection | undefined;
};

/**
 * Describes the message tilde.management.v1.UnassignCapabilityResponse.
 * Use `create(UnassignCapabilityResponseSchema)` to create a new message.
 */
export declare const UnassignCapabilityResponseSchema: GenMessage<UnassignCapabilityResponse>;

/**
 * @generated from message tilde.management.v1.ReconnectRequest
 */
export declare type ReconnectRequest = Message<"tilde.management.v1.ReconnectRequest"> & {
  /**
   * @generated from field: string id = 1;
   */
  id: string;
};

/**
 * Describes the message tilde.management.v1.ReconnectRequest.
 * Use `create(ReconnectRequestSchema)` to create a new message.
 */
export declare const ReconnectRequestSchema: GenMessage<ReconnectRequest>;

/**
 * @generated from message tilde.management.v1.ReconnectResponse
 */
export declare type ReconnectResponse = Message<"tilde.management.v1.ReconnectResponse"> & {
  /**
   * @generated from field: tilde.types.v1.Connection connection = 1;
   */
  connection?: Connection | undefined;

  /**
   * @generated from field: string brokering_url = 2;
   */
  brokeringUrl: string;
};

/**
 * Describes the message tilde.management.v1.ReconnectResponse.
 * Use `create(ReconnectResponseSchema)` to create a new message.
 */
export declare const ReconnectResponseSchema: GenMessage<ReconnectResponse>;

/**
 * @generated from message tilde.management.v1.DisconnectRequest
 */
export declare type DisconnectRequest = Message<"tilde.management.v1.DisconnectRequest"> & {
  /**
   * @generated from field: string id = 1;
   */
  id: string;
};

/**
 * Describes the message tilde.management.v1.DisconnectRequest.
 * Use `create(DisconnectRequestSchema)` to create a new message.
 */
export declare const DisconnectRequestSchema: GenMessage<DisconnectRequest>;

/**
 * @generated from message tilde.management.v1.DisconnectResponse
 */
export declare type DisconnectResponse = Message<"tilde.management.v1.DisconnectResponse"> & {
};

/**
 * Describes the message tilde.management.v1.DisconnectResponse.
 * Use `create(DisconnectResponseSchema)` to create a new message.
 */
export declare const DisconnectResponseSchema: GenMessage<DisconnectResponse>;

/**
 * @generated from service tilde.management.v1.ConnectionsService
 */
export declare const ConnectionsService: GenService<{
  /**
   * @generated from rpc tilde.management.v1.ConnectionsService.ListProviders
   */
  listProviders: {
    methodKind: "unary";
    input: typeof ListProvidersRequestSchema;
    output: typeof ListProvidersResponseSchema;
  },
  /**
   * @generated from rpc tilde.management.v1.ConnectionsService.GetProvider
   */
  getProvider: {
    methodKind: "unary";
    input: typeof GetProviderRequestSchema;
    output: typeof GetProviderResponseSchema;
  },
  /**
   * @generated from rpc tilde.management.v1.ConnectionsService.RegisterProvider
   */
  registerProvider: {
    methodKind: "unary";
    input: typeof RegisterProviderRequestSchema;
    output: typeof RegisterProviderResponseSchema;
  },
  /**
   * @generated from rpc tilde.management.v1.ConnectionsService.StartConnection
   */
  startConnection: {
    methodKind: "unary";
    input: typeof StartConnectionRequestSchema;
    output: typeof StartConnectionResponseSchema;
  },
  /**
   * @generated from rpc tilde.management.v1.ConnectionsService.GetConnection
   */
  getConnection: {
    methodKind: "unary";
    input: typeof GetConnectionRequestSchema;
    output: typeof GetConnectionResponseSchema;
  },
  /**
   * @generated from rpc tilde.management.v1.ConnectionsService.ListConnections
   */
  listConnections: {
    methodKind: "unary";
    input: typeof ListConnectionsRequestSchema;
    output: typeof ListConnectionsResponseSchema;
  },
  /**
   * @generated from rpc tilde.management.v1.ConnectionsService.AssignCapability
   */
  assignCapability: {
    methodKind: "unary";
    input: typeof AssignCapabilityRequestSchema;
    output: typeof AssignCapabilityResponseSchema;
  },
  /**
   * @generated from rpc tilde.management.v1.ConnectionsService.UnassignCapability
   */
  unassignCapability: {
    methodKind: "unary";
    input: typeof UnassignCapabilityRequestSchema;
    output: typeof UnassignCapabilityResponseSchema;
  },
  /**
   * @generated from rpc tilde.management.v1.ConnectionsService.Reconnect
   */
  reconnect: {
    methodKind: "unary";
    input: typeof ReconnectRequestSchema;
    output: typeof ReconnectResponseSchema;
  },
  /**
   * @generated from rpc tilde.management.v1.ConnectionsService.Disconnect
   */
  disconnect: {
    methodKind: "unary";
    input: typeof DisconnectRequestSchema;
    output: typeof DisconnectResponseSchema;
  },
}>;